requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
langchain==0.0.325
pdfminer.six==20221105
pdf2image==1.16.3
openai==0.28.1
gunicorn==20.1.0
tiktoken==0.5.1
faiss-cpu==1.7.4
